move periodic funcs to unified module
