--- src/periodic_functions/sine.rs
@@ -1,116 +1,0 @@
-//! Sine related functions.
-
-use crate::PeriodicFunction;
-use alloc::boxed::Box;
-use core::f64::consts::PI;
-
-#[cfg(all(not(feature = "libm"), feature = "std"))]
-fn _sine(frequency: f64, amplitude: f64, phase: f64) -> PeriodicFunction {
-    Box::new(move |t| {
-        let radians = (2.0 * PI * frequency * t) + (phase * 2.0 * PI);
-        let sine = radians.sin();
-
-        sine * amplitude
-    })
-}
-
-#[cfg(feature = "libm")]
-fn _sine(frequency: f64, amplitude: f64, phase: f64) -> PeriodicFunction {
-    use libm::sin;
-    Box::new(move |t| sin((2.0 * PI * frequency * t) + (phase * 2.0 * PI)) * amplitude)
-}
-
-/// Sine function builder. See the [`macro`] for more info.
-///
-/// [`macro`]: ../../macro.sine.html
-pub fn sine(frequency: f64, amplitude: f64, phase: f64) -> PeriodicFunction {
-    _sine(frequency, amplitude, phase)
-}
-
-/// Builder macro for Sine [PeriodicFunction].
-///
-/// Takes up to 3 arguments - frequency {amplitude, {phase}}
-///
-/// | argument | unit | notes |
-/// | -------- | ---- | ----- |
-/// | frequency | Hz | Frequecy of the periodic function. Also: 1 / period |
-/// | amplitude | *arbitrary* | The amplitude of the function in 0-peak notation. |
-/// | phase | *periods* | The phase shift of the function. Value of 1 means full shift around.
-///
-/// # Examples
-///
-/// 50 Hz sine of amplitude 1 and no phase shift
-/// ```
-/// use wavegen::sine;
-///
-/// let sine = sine!(50);
-/// ```
-///
-/// 50 Hz sine of amplitude 20 and no phase shift
-/// ```
-/// use wavegen::sine;
-///
-/// let sine = sine!(frequency: 50, amplitude: 20);
-/// ```
-///
-/// 50 Hz sine of amplitude 20 and phase shift of half a turn
-/// ```
-/// use core::f64::consts::PI;
-/// use wavegen::sine;
-///
-/// let sine = sine!(50, 20, 0.5);
-/// ```
-#[macro_export]
-macro_rules! sine {
-    (frequency: $frequency:expr) => {
-        sine!($frequency)
-    };
-    (frequency: $frequency:expr, amplitude: $amplitude:expr) => {
-        sine!($frequency, $amplitude)
-    };
-    (frequency: $frequency:expr, amplitude: $amplitude:expr, phase: $phase:expr) => {
-        sine!($frequency, $amplitude, $phase)
-    };
-    ($frequency:expr) => {
-        sine!($frequency, 1.0, 0.0)
-    };
-    ($frequency:expr, $amplitude:expr) => {
-        sine!($frequency, $amplitude, 0.0)
-    };
-    ($frequency:expr, $amplitude:expr, $phase:expr) => {
-        $crate::periodic_functions::sine::sine($frequency as f64, $amplitude as f64, $phase as f64)
-    };
-}
-
-#[cfg(test)]
-mod tests {
-    use float_cmp::approx_eq;
-
-    const EPS: f64 = 1e-3;
-
-    #[test]
-    fn default_sine_has_amplitude_of_one_and_no_phase_shift() {
-        let sine = sine!(1);
-
-        let max = sine(0.25);
-        let min = sine(0.75);
-        let zero = sine(0.5);
-
-        assert!(approx_eq!(f64, max, 1.0, epsilon = EPS));
-        assert!(approx_eq!(f64, min, -1.0, epsilon = EPS));
-        assert!(approx_eq!(f64, zero, 0.0, epsilon = EPS));
-    }
-
-    #[test]
-    fn phase_affects_min_max_amplitude_position() {
-        let sine = sine!(1, 1, 0.5);
-
-        let max = sine(0.75);
-        let min = sine(0.25);
-        let zero = sine(0.5);
-
-        assert!(approx_eq!(f64, max, 1.0, epsilon = EPS));
-        assert!(approx_eq!(f64, min, -1.0, epsilon = EPS));
-        assert!(approx_eq!(f64, zero, 0.0, epsilon = EPS));
-    }
-}
